// libs/features/kalite-kontrol/src/lib/kk-8246/enums/kk-8246-resource-url.ts
export enum Kk8246ResourceURL {
  TESISLER_ASITLEME = '/tesisler/asitleme',
  BOBIN_LISTESI = '/bobinler',
  TESISLER_GENEL = '/tesisler/genel',
  BAGIMSIZ_NUMUNE_BILGILERI = '/bagimsiz-numune-bilgileri',
  BAGIMSIZ_MESAJ_BILGILERI = '/bagimsiz-mesaj-bilgileri',
  PROSES_BILGISI_ASITLEME = '/proses-bilgisi-asitlemeler',
  PROSES_BILGISI_DURULAMA = '/proses-bilgisi-durulamalar',
  PROSES_BILGISI_YAGLAMA = '/proses-bilgisi-yaglamalar',
}